multi: free Rescan struct from ChainService dependency
In this commit, we free the Rescan struct from its ChainService
dependency. This is the start of an ongoing effort to better
abstract dependencies between different packages/functions, as the
ChainService has simply grown too large.

In order to achieve this, we introduce a ChainSource interface that will
be used by the Rescan struct. This interface is responsible for
retrieving information about the chain, such as, the chain parameters,
block headers, filter headers, etc.

This interface also includes a new Subscribe method, which allows users
of this interface to subscribe for new block notifications using the
recently introduced blockntfns package. Notifications are streamed
through a single channel, rather than two, in order to ensure that
notifications are processed in their correct order. As a result, we can
use this to simplify and abstract some of the logic within the rescan
method, e.g. waiting for the chain to be synced to the rescan's starting
point.