Remove dependency on NodeBloom
Previously we were using NodeBloom to provide a programatic toggle to turn on
and off receiving unconfirmed transactions. The alternative is a using the
relay bool in the version message but this requires a full restart to make
the change take affect.

However, the new interface in the bchwallet mobile package is very thin and
does not provide a method to toggle relay. So given this it makes sense to just
remove the dependency on NodeBloom so that remote peers do not need to support
NodeBloom in addition to NodeCF.