pushtx: introduce pushtx package with Broadcaster
In this commit, we introduce a new package: pushtx. This package will be
responsible for reliably propagating transactions throughout the
network. This is done by continuously broadcasting unconfirmed
transactions to peers upon every new update at the tip of the chain.
We'll continue doing this until we receive an acknowledgment from _all_
of our peers that the transaction has confirmed on-chain. This implies
that our trust model should include one well-behaved peer to ensure
transactions actually propagate throughout the network.

A new set of errors have also been introduced in order to introduce more
granularity to external callers. This is done by analyzing the contents
of a peer's reject message for a transaction we're attempting to
broadcast. Currently, four different cases exist:

  1. A transaction is deemed by a peer as invalid.
  2. A transaction is deemed by a peer as having an insufficient fee.
  3. A transaction is deemed by a peer as already being in its mempool.
  4. A transaction is deemed by a peer as already being included in the
     chain.