Change query timeout handling
This commit changes the handling of a timeout in the queryChainServiceBatch
function. Instead of existing the loop we'll remove our request from the list
and continue to the next iteration of the loop. This should be enough
to prevent getting into a bad state if a request comes back after timeout.

This is some extremely racey code and it seems like there's still a small
possibility of a race condition. Somthing to look at later on. But this change
is needed in case of accidental timeouts we don't want to stop querying a
good peer.

Ideally we'd be able to detect a bad peer that times out a lot and eventually
remove them from the query as they might slow the query down. But that is also
something for a later date.

In a perfect world the filters would be committed to the blocks and none of
this code would be needed and the entire neutrino codebase would be much more
simple.