banman: add Store interface with implementation
In this commit, we introduce a new Store interface, along with a
concrete implementation that will be backed by a boltdb instance, that
will serve as a persistent ban store for peers. This prevents us from
reconnecting to known bad peers until their ban expires.

The store has been designed so that it understands IP networks, which
give us greater flexibility as they can coalesce multiple IP addresses
that are likely to be correlated.