banman: add utility function ParseIPNet
We do this to facilitate the use of mapping peer addresses, which are
likely to be of the form host:port, into IP networks. An optional mask
can be provided, otherwise the default is used, which is currently set
so that the IP network _only_ contains the specified peer address.