headerfs: add ability to assert the state of the filter header chain
In this commit, we add the ability to specify an expected filter header
state when the filter header store is being created. As the filter
headers aren't committed yet, it's at times useful to nudge existing
nodes towards a filter header chain that they themselves may not have
obtained previously.