query: use per-query matchSignals.
This fixes an issue that could happen if a response to a previously
timed out query was received during sync. This could cause the response
handler to send on the peer's matchSignal, but the peerGoroutine was no
longer listening on the channel, causing sync to stall.

Instead we tie the matchSignal channel to the specific query, such that
we know that we only send the signal if the current query is
successfully handled, and we buffer the channel to ensure no blocking
even after a timeout.