query: reset the main timeout each time a message is received
In this commit, we aim to fix an issue that can cause a query to fail
all together for a large query over a possibly high latency link. Before
this commit, we if attempted to query for several filters at once, on
mobile phones, it's possible that we timeout part way through. As is,
we'll then exit, and fail to properly retry, messing up the main rescan
loop.

In this commit, we aim to address this by _resetting_ the `peerTimeout`
each time we receive a message. Rather than aim to complete the entire
query in the timeout interval, we instead modify the timeout to ensure
forward progress of the query. If the peer isn't responding, then the
`peerTimeout` will tick, and we'll switch to another peer.

As a result, it's now possible that a query takes longer than
`QueryPeerConnectTimeout` if a peer is slow, or we're using a high
latency link, or have a meager CPU. As a result, we'll now _stop_ the
now named `connectionTimeout` once we receive the first message from the
peer. If the peer stops responding all together, then the `peerTimeout`
will be ticked. Arguably, we don't even need the `connectionTimout` due
to this new logic.

Fixes #164.